add meetup group details

+++ Models/Domain/MeetupCity.cs
@@ -9,7 +9,7 @@ namespace DataDrivenDiversity.Models.Domain
         public string City { get; set; }// "Beerwah",
         public float Lon { get; set; }// 152.9600067138672,
         public int Ranking { get; set; }// 19,
-        public double Id { get; set; }// 1111057,
+        public long Id { get; set; }// 1111057,
         public int MemberCount { get; set; }// 0,
         public float Lat { get; set; }// -26.860000610351562
     }
